test: add shared API Gateway helpers and VPC-style WebsocketApi settlement
