+++ tests/aws/api_gateway/http_api/test_http_api_domain.py
@@ -10,93 +10,11 @@ from pytest import mark, raises
 from stelvio.aws.api_gateway import ApiDomain, HttpApi, HttpApiConfig
 from stelvio.dns import DnsProviderNotConfiguredError
 
-from ...pulumi_mocks import ACCOUNT_ID, DEFAULT_REGION, R, tid
+from ...pulumi_mocks import R, tid
+from ..conftest import API_DOMAIN_GRAPH_COUNTS, assert_api_domain_graph
 from .conftest import TP, when_http_api_ready
 
 pytestmark = mark.usefixtures("project_cwd")
-
-_DOMAIN_GRAPH_COUNTS = {
-    R.CERTIFICATE: 1,
-    R.CLOUDFLARE_RECORD: 2,
-    R.CERTIFICATE_VALIDATION: 1,
-    R.HTTP_API_DOMAIN_NAME: 1,
-}
-
-
-def _certificate_arn(domain_component: str) -> str:
-    return (
-        f"arn:aws:acm:{DEFAULT_REGION}:{ACCOUNT_ID}:certificate/"
-        f"{tid(TP + f'{domain_component}-cert-certificate')}"
-    )
-
-
-def assert_api_domain_graph(
-    mocks,
-    *,
-    domain_component: str,
-    domain_name: str,
-    domain_extra_inputs: dict[str, Any] | None = None,
-    dns_record_extra_inputs: dict[str, Any] | None = None,
-) -> None:
-    """Assert ACM, DomainName, and DNS resources for an ApiDomain."""
-    certificate_arn = _certificate_arn(domain_component)
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate",
-        R.CERTIFICATE,
-        {"domainName": domain_name, "validationMethod": "DNS"},
-    )
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate-validation",
-        R.CERTIFICATE_VALIDATION,
-        {
-            "certificateArn": certificate_arn,
-            "validationRecordFqdns": [f"_test.{domain_name}"],
-        },
-    )
-    domain_inputs: dict[str, Any] = {
-        "domainName": domain_name,
-        "domainNameConfiguration": {
-            "certificateArn": certificate_arn,
-            "endpointType": "REGIONAL",
-            "securityPolicy": "TLS_1_2",
-        },
-    }
-    if domain_extra_inputs:
-        domain_inputs.update(domain_extra_inputs)
-    mocks.assert_res(
-        f"{domain_component}-domain",
-        R.HTTP_API_DOMAIN_NAME,
-        domain_inputs,
-        partial=domain_extra_inputs is not None,
-    )
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate-validation-record",
-        R.CLOUDFLARE_RECORD,
-        {
-            "name": f"_test.{domain_name}",
-            "type": "CNAME",
-            "content": f"test-validation.{domain_name}",
-            "ttl": 1.0,
-        },
-        partial=True,
-    )
-    dns_inputs: dict[str, Any] = {
-        "name": domain_name,
-        "type": "CNAME",
-        "content": (
-            f"d-{tid(TP + f'{domain_component}-domain')}"
-            f".execute-api.{DEFAULT_REGION}.amazonaws.com"
-        ),
-        "ttl": 300.0,
-    }
-    if dns_record_extra_inputs:
-        dns_inputs.update(dns_record_extra_inputs)
-    mocks.assert_res(
-        f"{domain_component}-dns-record",
-        R.CLOUDFLARE_RECORD,
-        dns_inputs,
-        partial=True,
-    )
 
 
 def assert_api_mapping(
@@ -135,7 +53,7 @@ def _http_api_with_domain_counts(
         R.LAMBDA_PERMISSION: function_count,
         R.HTTP_API_ROUTE: function_count,
         R.HTTP_API_MAPPING: mapping_count,
-        **_DOMAIN_GRAPH_COUNTS,
+        **API_DOMAIN_GRAPH_COUNTS,
     }
 
 
@@ -207,124 +125,6 @@ def test_http_api_config_conflicts_with_domain_option(app_context_with_dns):
 
     with raises(ValueError, match="cannot combine 'config' parameter"):
         HttpApi("my-api", config=HttpApiConfig(domain=config_domain), domain=keyword_domain)
-
-
-def _when_api_domain_ready(domain: ApiDomain, callback) -> None:
-    """Wait until the domain and its public DNS record are registered."""
-    resources = domain.resources
-    pulumi.Output.all(
-        resources.custom_domain.domain_name,
-        resources.custom_domain.domain_name_configuration,
-        resources.dns_record.pulumi_resource.id,
-    ).apply(callback)
-
-
-@pulumi.runtime.test
-def test_http_api_domain_dns_record_customize_applies_only_to_public_record(
-    pulumi_mocks,
-    app_context_with_dns,
-):
-    domain = ApiDomain(
-        "shared-domain",
-        domain_name="api.example.com",
-        customize={"dns_record": {"ttl": 600}},
-    )
-    _ = domain.resources
-
-    def check(_):
-        assert_api_domain_graph(
-            pulumi_mocks,
-            domain_component="shared-domain",
-            domain_name="api.example.com",
-            dns_record_extra_inputs={"ttl": 600.0},
-        )
-        # Public record ttl customized; validation record remains ttl 1.
-        validation = app_context_with_dns.created_records[0]
-        public = app_context_with_dns.created_records[1]
-        assert validation[4] == 1
-        assert public[4] == 600
-        pulumi_mocks.assert_res_counts(dict(_DOMAIN_GRAPH_COUNTS))
-
-    _when_api_domain_ready(domain, check)
-
-
-@pulumi.runtime.test
-def test_http_api_domain_customize_domain_key(pulumi_mocks, app_context_with_dns):
-    domain = ApiDomain(
-        "shared-domain",
-        domain_name="api.example.com",
-        customize={"domain": {"tags": {"Purpose": "test"}}},
-    )
-    _ = domain.resources
-
-    def check(_):
-        assert_api_domain_graph(
-            pulumi_mocks,
-            domain_component="shared-domain",
-            domain_name="api.example.com",
-            domain_extra_inputs={"tags": {"Purpose": "test"}},
-        )
-        pulumi_mocks.assert_res_counts(dict(_DOMAIN_GRAPH_COUNTS))
-
-    _when_api_domain_ready(domain, check)
-
-
-_EXTERNAL_CERT_ARN = f"arn:aws:acm:{DEFAULT_REGION}:{ACCOUNT_ID}:certificate/pre-provisioned"
-
-
-@pulumi.runtime.test
-def test_api_domain_uses_external_certificate_arn(pulumi_mocks, app_context_with_dns):
-    domain = ApiDomain(
-        "shared-domain",
-        domain_name="api.example.com",
-        certificate_arn=_EXTERNAL_CERT_ARN,
-    )
-    _ = domain.resources
-
-    def check(_):
-        assert domain.resources.acm_domain is None
-        pulumi_mocks.assert_res(
-            "shared-domain-domain",
-            R.HTTP_API_DOMAIN_NAME,
-            {
-                "domainName": "api.example.com",
-                "domainNameConfiguration": {
-                    "certificateArn": _EXTERNAL_CERT_ARN,
-                    "endpointType": "REGIONAL",
-                    "securityPolicy": "TLS_1_2",
-                },
-            },
-        )
-        pulumi_mocks.assert_res(
-            "shared-domain-dns-record",
-            R.CLOUDFLARE_RECORD,
-            {
-                "name": "api.example.com",
-                "type": "CNAME",
-                "ttl": 300.0,
-            },
-            partial=True,
-        )
-        pulumi_mocks.assert_res_counts(
-            {
-                R.HTTP_API_DOMAIN_NAME: 1,
-                R.CLOUDFLARE_RECORD: 1,
-            }
-        )
-
-    _when_api_domain_ready(domain, check)
-
-
-def test_api_domain_rejects_certificate_arn_with_certificate_customize(
-    app_context_with_dns,
-):
-    with raises(ValueError, match="Cannot specify both 'certificate_arn'"):
-        ApiDomain(
-            "shared-domain",
-            domain_name="api.example.com",
-            certificate_arn=_EXTERNAL_CERT_ARN,
-            customize={"certificate": {"tags": {"Env": "test"}}},
-        )
 
 
 def test_http_api_domain_requires_dns_provider(app_context_without_dns):
