+++ tests/aws/api_gateway/websocket_api/conftest.py
@@ -1,13 +1,6 @@
-import json
-from collections.abc import Sequence
 from typing import Any
 
-import pulumi
-
-from stelvio.aws.api_gateway import ApiDomain
-from stelvio.aws.api_gateway.websocket_api import WebsocketApi
-
-from ...pulumi_mocks import ACCOUNT_ID, DEFAULT_REGION, R, tid, tn
+from ...pulumi_mocks import ACCOUNT_ID, DEFAULT_REGION, R, tid
 
 TP = "test-test-"
 
@@ -17,169 +10,6 @@ LAMBDA_INVOKE_ARN_TEMPLATE = (
     f"arn:aws:apigateway:{DEFAULT_REGION}:lambda:path/2015-03-31/functions/"
     f"arn:aws:lambda:{DEFAULT_REGION}:{ACCOUNT_ID}:function:{{function_name}}/invocations"
 )
-LAMBDA_ASSUME_ROLE_POLICY = [
-    {
-        "actions": ["sts:AssumeRole"],
-        "principals": [{"identifiers": ["lambda.amazonaws.com"], "type": "Service"}],
-    }
-]
-
-_DOMAIN_GRAPH_COUNTS = {
-    R.CERTIFICATE: 1,
-    R.CLOUDFLARE_RECORD: 2,
-    R.CERTIFICATE_VALIDATION: 1,
-    R.HTTP_API_DOMAIN_NAME: 1,
-}
-
-
-def when_websocket_api_ready(api: WebsocketApi | Sequence[WebsocketApi], callback) -> None:
-    """Trigger callback after all WebSocket API resources are created.
-
-    Waits on api, stage, log group, and api mapping (when present)
-    so all resources are registered before assertions run. Accepts one
-    API or a sequence when multiple APIs must finish together.
-    """
-    apis = (api,) if isinstance(api, WebsocketApi) else api
-    outputs = []
-    for websocket_api in apis:
-        resources = websocket_api.resources
-        outputs.append(resources.api.id)
-        outputs.append(resources.stage.id)
-        outputs.append(resources.log_group.id)
-        if resources.api_mapping is not None:
-            outputs.append(resources.api_mapping.id)
-    pulumi.Output.all(*outputs).apply(callback)
-
-
-def when_api_domain_ready(domain: ApiDomain, callback) -> None:
-    """Wait until the domain and its public DNS record are registered."""
-    resources = domain.resources
-    pulumi.Output.all(
-        resources.custom_domain.domain_name,
-        resources.custom_domain.domain_name_configuration,
-        resources.dns_record.pulumi_resource.id,
-    ).apply(callback)
-
-
-def websocket_api_counts(  # noqa: PLR0913
-    *,
-    function_count: int = 0,
-    route_count: int = 0,
-    integration_count: int | None = None,
-    permission_count: int | None = None,
-    authorizer_count: int = 0,
-    mapping_count: int = 0,
-    with_domain: bool = False,
-    policy_count: int = 0,
-    extra: dict[R, int] | None = None,
-) -> dict[R, int]:
-    """Exact resource counts for a single WebsocketApi resource graph.
-
-    Always includes access-log wiring from ``_create_api_gateway_account_and_role``:
-    two ``API_ACCOUNT`` resources (``Account.get`` ref + conditional ``Account`` write)
-    and one CloudWatch push-to-logs role (+1 on ``ROLE``).
-    """
-    integration_count = function_count if integration_count is None else integration_count
-    permission_count = function_count if permission_count is None else permission_count
-    counts: dict[R, int] = {
-        R.HTTP_API: 1,
-        R.HTTP_API_STAGE: 1,
-        R.API_ACCOUNT: 2,  # Account.get ref + Account write from shared IAM helper
-        R.LOG_GROUP: 1,
-        R.ROLE: function_count + 1,  # +1 CloudWatch push-to-logs role
-    }
-    if function_count:
-        counts[R.FUNCTION] = function_count
-        counts[R.ROLE_POLICY_ATTACHMENT] = function_count + policy_count
-    if policy_count:
-        counts[R.POLICY] = policy_count
-    if integration_count:
-        counts[R.HTTP_API_INTEGRATION] = integration_count
-    if permission_count:
-        counts[R.LAMBDA_PERMISSION] = permission_count
-    if route_count:
-        counts[R.HTTP_API_ROUTE] = route_count
-    if authorizer_count:
-        counts[R.HTTP_API_AUTHORIZER] = authorizer_count
-    if mapping_count:
-        counts[R.HTTP_API_MAPPING] = mapping_count
-    if with_domain:
-        counts |= _DOMAIN_GRAPH_COUNTS
-    if extra:
-        counts |= extra
-    return counts
-
-
-def assert_api_domain_graph(
-    mocks,
-    *,
-    domain_component: str,
-    domain_name: str,
-    domain_extra_inputs: dict[str, Any] | None = None,
-    dns_record_extra_inputs: dict[str, Any] | None = None,
-) -> None:
-    """Assert ACM, DomainName, and DNS resources for an ApiDomain."""
-    certificate_arn = (
-        f"arn:aws:acm:{DEFAULT_REGION}:{ACCOUNT_ID}:certificate/"
-        f"{tid(TP + f'{domain_component}-cert-certificate')}"
-    )
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate",
-        R.CERTIFICATE,
-        {"domainName": domain_name, "validationMethod": "DNS"},
-    )
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate-validation",
-        R.CERTIFICATE_VALIDATION,
-        {
-            "certificateArn": certificate_arn,
-            "validationRecordFqdns": [f"_test.{domain_name}"],
-        },
-    )
-    domain_inputs: dict[str, Any] = {
-        "domainName": domain_name,
-        "domainNameConfiguration": {
-            "certificateArn": certificate_arn,
-            "endpointType": "REGIONAL",
-            "securityPolicy": "TLS_1_2",
-        },
-    }
-    if domain_extra_inputs:
-        domain_inputs.update(domain_extra_inputs)
-    mocks.assert_res(
-        f"{domain_component}-domain",
-        R.HTTP_API_DOMAIN_NAME,
-        domain_inputs,
-        partial=domain_extra_inputs is not None,
-    )
-    mocks.assert_res(
-        f"{domain_component}-cert-certificate-validation-record",
-        R.CLOUDFLARE_RECORD,
-        {
-            "name": f"_test.{domain_name}",
-            "type": "CNAME",
-            "content": f"test-validation.{domain_name}",
-            "ttl": 1.0,
-            "zoneId": "test-zone-id",
-        },
-    )
-    dns_inputs: dict[str, Any] = {
-        "name": domain_name,
-        "type": "CNAME",
-        "content": (
-            f"d-{tid(TP + f'{domain_component}-domain')}"
-            f".execute-api.{DEFAULT_REGION}.amazonaws.com"
-        ),
-        "ttl": 300.0,
-        "zoneId": "test-zone-id",
-    }
-    if dns_record_extra_inputs:
-        dns_inputs.update(dns_record_extra_inputs)
-    mocks.assert_res(
-        f"{domain_component}-dns-record",
-        R.CLOUDFLARE_RECORD,
-        dns_inputs,
-    )
 
 
 def assert_api_mapping(
@@ -198,20 +28,3 @@ def assert_api_mapping(
     if mapping_key is not None:
         inputs["apiMappingKey"] = mapping_key
     mocks.assert_res(f"{api_name}-api-mapping", R.HTTP_API_MAPPING, inputs)
-
-
-def assert_lambda_role_and_attachment(mocks, function_name: str) -> None:
-    role_name = f"{function_name}-r"
-    mocks.assert_res(
-        role_name,
-        R.ROLE,
-        {"assumeRolePolicy": json.dumps(LAMBDA_ASSUME_ROLE_POLICY)},
-    )
-    mocks.assert_res(
-        f"{function_name}-basic-execution-r-p-attachment",
-        R.ROLE_POLICY_ATTACHMENT,
-        {
-            "policyArn": "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole",
-            "role": tn(TP + role_name),
-        },
-    )
