+++ tests/aws/api_gateway/websocket_api/test_websocket_api_routes.py
@@ -7,7 +7,7 @@ from stelvio.aws.api_gateway.websocket_api import WebsocketApi
 from stelvio.aws.function import Function, FunctionConfig
 
 from ...pulumi_mocks import ACCOUNT_ID, DEFAULT_REGION, R, tid, tn
-from .conftest import TP, WEBSOCKET_API_ID, websocket_api_counts, when_websocket_api_ready
+from .conftest import TP, WEBSOCKET_API_ID
 
 pytestmark = mark.usefixtures("project_cwd")
 
@@ -38,72 +38,86 @@ def test_websocket_api_rejects_duplicate_routes():
     ],
     ids=["string_handler_and_opts", "function_config", "dict"],
 )
-@pulumi.runtime.test
 def test_websocket_api_route_handler_configuration(pulumi_mocks, handler, opts):
     api = WebsocketApi("chat")
     api.route("$connect", handler, **opts)
-    _ = api.resources
 
-    def check(_):
-        # Lambda Function inputs include code/role assets — assert the configured fields.
-        pulumi_mocks.assert_res(
-            "chat-functions-simple_handler",
-            R.FUNCTION,
-            {"handler": "simple.handler", "memorySize": 512.0, "timeout": 60.0},
-            partial=True,
-        )
-        pulumi_mocks.assert_res_counts(
-            websocket_api_counts(
-                function_count=1,
-                route_count=1,
-                integration_count=1,
-                permission_count=1,
-            )
-        )
+    @pulumi.runtime.test
+    def deploy():
+        return api.resources
 
-    when_websocket_api_ready(api, check)
+    deploy()
+
+    # Lambda Function inputs include code/role assets — assert the configured fields.
+    pulumi_mocks.assert_res(
+        "chat-functions-simple_handler",
+        R.FUNCTION,
+        {"handler": "simple.handler", "memorySize": 512.0, "timeout": 60.0},
+        partial=True,
+    )
+    pulumi_mocks.assert_res_counts(
+        {
+            R.HTTP_API: 1,
+            R.HTTP_API_STAGE: 1,
+            R.API_ACCOUNT: 2,
+            R.LOG_GROUP: 1,
+            R.ROLE: 2,
+            R.FUNCTION: 1,
+            R.ROLE_POLICY_ATTACHMENT: 1,
+            R.HTTP_API_INTEGRATION: 1,
+            R.LAMBDA_PERMISSION: 1,
+            R.HTTP_API_ROUTE: 1,
+        }
+    )
 
 
-@pulumi.runtime.test
 def test_websocket_api_route_uses_supplied_function(pulumi_mocks):
     function = Function("connect", handler="functions/simple.handler")
     api = WebsocketApi("chat")
     api.route("$connect", function)
-    _ = api.resources
 
-    def check(_):
-        pulumi_mocks.assert_res("connect", R.FUNCTION)
-        pulumi_mocks.assert_res(
-            "chat-permission-chat-connect",
-            R.LAMBDA_PERMISSION,
-            {
-                "action": "lambda:InvokeFunction",
-                "function": tn(TP + "connect"),
-                "principal": "apigateway.amazonaws.com",
-                "sourceArn": (
-                    f"arn:aws:execute-api:{DEFAULT_REGION}:{ACCOUNT_ID}:{WEBSOCKET_API_ID}/*/*"
-                ),
-            },
-        )
-        pulumi_mocks.assert_res(
-            "chat-route-sys-connect",
-            R.HTTP_API_ROUTE,
-            {
-                "apiId": WEBSOCKET_API_ID,
-                "routeKey": "$connect",
-                "target": f"integrations/{tid(TP + 'chat-integration-chat-connect')}",
-            },
-        )
-        pulumi_mocks.assert_res_counts(
-            websocket_api_counts(
-                function_count=1,
-                route_count=1,
-                integration_count=1,
-                permission_count=1,
-            )
-        )
+    @pulumi.runtime.test
+    def deploy():
+        return api.resources
 
-    when_websocket_api_ready(api, check)
+    deploy()
+
+    pulumi_mocks.assert_res("connect", R.FUNCTION)
+    pulumi_mocks.assert_res(
+        "chat-permission-chat-connect",
+        R.LAMBDA_PERMISSION,
+        {
+            "action": "lambda:InvokeFunction",
+            "function": tn(TP + "connect"),
+            "principal": "apigateway.amazonaws.com",
+            "sourceArn": (
+                f"arn:aws:execute-api:{DEFAULT_REGION}:{ACCOUNT_ID}:{WEBSOCKET_API_ID}/*/*"
+            ),
+        },
+    )
+    pulumi_mocks.assert_res(
+        "chat-route-sys-connect",
+        R.HTTP_API_ROUTE,
+        {
+            "apiId": WEBSOCKET_API_ID,
+            "routeKey": "$connect",
+            "target": f"integrations/{tid(TP + 'chat-integration-chat-connect')}",
+        },
+    )
+    pulumi_mocks.assert_res_counts(
+        {
+            R.HTTP_API: 1,
+            R.HTTP_API_STAGE: 1,
+            R.API_ACCOUNT: 2,
+            R.LOG_GROUP: 1,
+            R.ROLE: 2,
+            R.FUNCTION: 1,
+            R.ROLE_POLICY_ATTACHMENT: 1,
+            R.HTTP_API_INTEGRATION: 1,
+            R.LAMBDA_PERMISSION: 1,
+            R.HTTP_API_ROUTE: 1,
+        }
+    )
 
 
 @mark.parametrize(
@@ -183,22 +197,29 @@ def test_websocket_api_conflicting_lambda_configurations_raise(pulumi_mocks, fir
         _ = api.resources
 
 
-@pulumi.runtime.test
 def test_websocket_api_folder_handlers_get_distinct_lambdas(pulumi_mocks):
     """folder/:: configs that share a handler suffix must not collide."""
     api = WebsocketApi("chat")
     api.route("$connect", "functions/folder::handler.fn")
     api.route("$disconnect", "functions/folder2::handler.fn")
-    _ = api.resources
 
-    def check(_):
-        pulumi_mocks.assert_res_counts(
-            websocket_api_counts(
-                function_count=2,
-                route_count=2,
-                integration_count=2,
-                permission_count=2,
-            )
-        )
+    @pulumi.runtime.test
+    def deploy():
+        return api.resources
 
-    when_websocket_api_ready(api, check)
+    deploy()
+
+    pulumi_mocks.assert_res_counts(
+        {
+            R.HTTP_API: 1,
+            R.HTTP_API_STAGE: 1,
+            R.API_ACCOUNT: 2,
+            R.LOG_GROUP: 1,
+            R.ROLE: 3,
+            R.FUNCTION: 2,
+            R.ROLE_POLICY_ATTACHMENT: 2,
+            R.HTTP_API_INTEGRATION: 2,
+            R.LAMBDA_PERMISSION: 2,
+            R.HTTP_API_ROUTE: 2,
+        }
+    )
